static_cast to double for pow
Windows (incorrectly) provides overloads for pow so that if T is not
double/float it won't compile

+++ tests/arithmetics.cpp
@@ -353,7 +353,7 @@ template<typename Vec> void testProduct()
                     x2 *= x;
                 }
             } else {
-                x2 = pow(x, m.count());
+                x2 = pow(static_cast<double>(x), m.count());
             }
             COMPARE(v.product(m), x2) << m << v;
         } while (true);
